Transform translate activity to fragment
