+++ app/src/main/java/com/paulacr/wordslearning/feature/wordslist/WordsListFragment.kt
@@ -1,0 +1,8 @@
+package com.paulacr.wordslearning.feature.wordslist
+
+import androidx.fragment.app.Fragment
+
+class WordsListFragment : Fragment() {
+
+
+}
